backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
requests>=2.31.0